Rope refactor: cli -> main
